packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-test-observability"
version = "0.1.0"
description = "Trace capture, tool call logs, reasoning timings, and latency breakdowns"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["test-observability","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-test-observability"
Repository = "https://github.com/mxn2020/minions-test-observability"

[tool.hatch.build.targets.wheel]
packages = ["minions_test_observability"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
